Separate db code into separate file
